+++ CHANGELOG.md
@@ -26,6 +26,10 @@
 * Added `WriteString()`
 * Added `closed` flag
 
+### Refactoring
+
+* `TextBuffer` and `TextFile` now use common code.
+
 ## v2.1.1
 
 Released Wednesday, 24th November 2021.